web-95-sidebar-cards: create new sidebar cards
