+++ wetmap/src/components/boundaryDiveShops/diveShopItem/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import getPhotoPublicUrl from '../../../helpers/getPhotoPublicUrl';
 import { DiveShop } from '../../../entities/diveShop';
-import ListItemWithImage from '../../reusables/listItemWithImage';
+import SidebarCard from '../../reusables/sidebarCard';
 
 type DiveShopItemProps = {
   diveShop: DiveShop
@@ -11,10 +11,9 @@ export default function DiveShopItem(props: DiveShopItemProps) {
   const imageUrl = getPhotoPublicUrl(props.diveShop.diveshopprofilephoto);
 
   return (
-    <ListItemWithImage
+    <SidebarCard
       imageUrl={imageUrl}
       title={props.diveShop.orgname}
-      date={props.diveShop.diveshopbio}
     />
   );
 }
