+++ wetmap/src/components/reusables/sidebarCard/style.module.scss
@@ -1,0 +1,74 @@
+.card {
+  position: relative;
+  z-index: 2;
+  width: 100%;
+  height: 12rem;
+  min-height: 12rem;
+  border-radius: 1rem;
+  overflow: hidden;
+  color: white;
+  padding: 1.25rem;
+  display: flex;
+  flex-direction: column;
+  justify-content: end;
+  align-items: start;
+  gap: 0.25rem;
+  cursor: pointer;
+  border: 2px solid white;
+  outline: 2px solid white;
+}
+
+.highlighted {
+  border: 2px solid white;
+  outline: 2px solid #0073E6;
+}
+
+.title {
+  margin: 0;
+  font-weight: 500;
+  font-size: 18px;
+}
+
+.backgroundImage {
+  position: absolute;
+  z-index: -2;
+  inset: 0;
+  width: 100%;
+  height: 100%;
+  background-color: darkgray;
+  object-fit: cover;
+  object-position: 50% 50%;
+  transition: 250ms ease;
+}
+
+.card:hover .backgroundImage {
+  transform: scale(1.05);
+}
+
+.info {
+  font-size: 14px;
+  display: flex;
+  align-items: center;
+  gap: 0.25rem;
+  opacity: 0.8;
+}
+
+.rating svg {
+  width: 0.75rem;
+  height: auto;
+}
+
+.rating {
+  display: flex;
+  align-items: center;
+  gap: 0.2rem;
+}
+
+.overlay {
+  position: absolute;
+  z-index: -1;
+  inset: 0;
+  width: 100%;
+  height: 100%;
+  background: linear-gradient(10deg, rgba(0, 0, 0) 0%, rgba(0, 0, 0, 0.75) 35%, rgba(0, 0, 0, 0) 75%);;
+}
